GLSDK BACKPORT: km: fix GEM handle memory leak
Backport of commit 0f3561a47cf7a1b968fb64446a0be10abc9a15d5 from
omap5-sgx-ddk-linux.

When memory handles are mapped, a dummy GEM handle is created so that it can be
accepted by DRM. However, when the memory handle is unmaped, there is no
corresponding delete of  the GEM handle.

Given that each GEM handle is 4K bytes, this will result in a memory leak that
will result in application crash over time. This is not catastrophic to the
system since DRM/GEM core will cleanup the GEM handles allocated as part of the
proces when it exit.

This patch addresses the GEM handle leak in SGX driver.

Signed-off-by: Anand Balagopalakrishnan <anandb@ti.com>
[tony@atomide.com: updated to apply]
Signed-off-by: Tony Lindgren <tony@atomide.com>