drm/omap: Remove omap_gem_set_sync_object and waiter->read/write_complete
We have PVRSRV_DISABLE_UM_SYNCOBJ_MAPPINGS defined in pvr-omap4-dkms. This
means that we can simplify things by getting rid of the tinkering of the
sync object from pvr driver.

This means we could make the sync code just use the standard refcount API.

And we are just copying values to struct omap_gem_sync_waiter and we
do it without spinlock meaning the value may not be up to date.

Let's just get rid of the duplicate data in omap_gem_sync_waiter.

Signed-off-by: Tony Lindgren <tony@atomide.com>